chore(email-security): update generated types and methods

+++ src/resources/email-security/settings/trusted-domains.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import { V4PagePaginationArray, type V4PagePaginationArrayParams } from '../../../pagination';
 
@@ -27,7 +28,7 @@ export class TrustedDomains extends APIResource {
     params: TrustedDomainCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<TrustedDomainCreateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/email-security/settings/trusted_domains`, {
         body,
@@ -50,10 +51,20 @@ export class TrustedDomains extends APIResource {
    * ```
    */
   list(
-    params: TrustedDomainListParams,
+    params?: TrustedDomainListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<TrustedDomainListResponsesV4PagePaginationArray, TrustedDomainListResponse>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<TrustedDomainListResponsesV4PagePaginationArray, TrustedDomainListResponse>;
+  list(
+    params: TrustedDomainListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<TrustedDomainListResponsesV4PagePaginationArray, TrustedDomainListResponse> {
-    const { account_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId, ...query } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/email-security/settings/trusted_domains`,
       TrustedDomainListResponsesV4PagePaginationArray,
@@ -76,10 +87,22 @@ export class TrustedDomains extends APIResource {
    */
   delete(
     trustedDomainId: number,
-    params: TrustedDomainDeleteParams,
+    params?: TrustedDomainDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<TrustedDomainDeleteResponse>;
+  delete(
+    trustedDomainId: number,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<TrustedDomainDeleteResponse>;
+  delete(
+    trustedDomainId: number,
+    params: TrustedDomainDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<TrustedDomainDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(trustedDomainId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(
         `/accounts/${account_id}/email-security/settings/trusted_domains/${trustedDomainId}`,
@@ -105,7 +128,7 @@ export class TrustedDomains extends APIResource {
     params: TrustedDomainEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<TrustedDomainEditResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(
         `/accounts/${account_id}/email-security/settings/trusted_domains/${trustedDomainId}`,
@@ -128,10 +151,19 @@ export class TrustedDomains extends APIResource {
    */
   get(
     trustedDomainId: number,
-    params: TrustedDomainGetParams,
+    params?: TrustedDomainGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<TrustedDomainGetResponse>;
+  get(trustedDomainId: number, options?: Core.RequestOptions): Core.APIPromise<TrustedDomainGetResponse>;
+  get(
+    trustedDomainId: number,
+    params: TrustedDomainGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<TrustedDomainGetResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(trustedDomainId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(
         `/accounts/${account_id}/email-security/settings/trusted_domains/${trustedDomainId}`,
@@ -310,7 +342,7 @@ export declare namespace TrustedDomainCreateParams {
     /**
      * Path param: Account Identifier
      */
-    account_id: string;
+    account_id?: string;
 
     /**
      * Body param: Select to prevent recently registered domains from triggering a
@@ -345,7 +377,7 @@ export declare namespace TrustedDomainCreateParams {
     /**
      * Path param: Account Identifier
      */
-    account_id: string;
+    account_id?: string;
 
     /**
      * Body param
@@ -380,7 +412,7 @@ export interface TrustedDomainListParams extends V4PagePaginationArrayParams {
   /**
    * Path param: Account Identifier
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Query param: The sorting direction.
@@ -419,14 +451,14 @@ export interface TrustedDomainDeleteParams {
   /**
    * Account Identifier
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface TrustedDomainEditParams {
   /**
    * Path param: Account Identifier
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -461,7 +493,7 @@ export interface TrustedDomainGetParams {
   /**
    * Account Identifier
    */
-  account_id: string;
+  account_id?: string;
 }
 
 TrustedDomains.TrustedDomainListResponsesV4PagePaginationArray =
